Make GraphBuilder OS Agnostic

+++ SharpViz/GraphBuilder.cs
@@ -8,17 +8,17 @@ namespace SharpViz
     {
         public static void GenAndOpenDotGraph(string dotSyntax)
         {
-            GenAndOpen(dotSyntax, "dot.exe");
+            GenAndOpen(dotSyntax, "dot");
         }
 
         public static void GenAndOpenForceDirectedGraph(string dotSyntax)
         {
-            GenAndOpen(dotSyntax, "neato.exe");
+            GenAndOpen(dotSyntax, "neato");
         }
 
         public static string GenerateDotGraph(string dotSyntax)
         {
-            return GeneratePdf(dotSyntax, "dot.exe");
+            return GeneratePdf(dotSyntax, "dot");
         }
 
         private static void GenAndOpen(string dotSyntax, string executable)
@@ -32,9 +32,11 @@ namespace SharpViz
 
             var fileOpenProcess = new Process
             {
-                StartInfo = new ProcessStartInfo(outFile)
+                StartInfo = new ProcessStartInfo()
                 {
-                    UseShellExecute = true
+                    UseShellExecute = true,
+                    FileName = OS() == PlatformID.Unix || OS() == PlatformID.MacOSX ? "open" : outFile,
+                    Arguments = OS() == PlatformID.Unix || OS() == PlatformID.MacOSX ? outFile : ""
                 }
             };
             fileOpenProcess.Start();
@@ -42,10 +44,15 @@ namespace SharpViz
 
         private static string GeneratePdf(string dotSyntax, string executable)
         {
-            var dotFile = Path.GetTempFileName();
-            var outFile = Path.GetTempFileName();
-            File.Delete(outFile);
-            outFile += ".pdf";
+            var dotFile = Path.GetTempPath() + Path.GetRandomFileName() + ".dot";
+            var outFile = Path.GetTempPath() + Path.GetRandomFileName() + ".pdf";
+            var execPath = GetFilePath(executable);
+
+            if(!File.Exists(execPath)) {
+                Console.Error.Write("{0} does not exist. Please make sure Graphviz is installed", execPath);
+                return null;
+            }
+
             File.WriteAllText(dotFile, dotSyntax);
 
             var process = new Process();
@@ -53,7 +60,7 @@ namespace SharpViz
             process.StartInfo.RedirectStandardError = true;
             process.StartInfo.UseShellExecute = false;
             process.StartInfo.CreateNoWindow = true;
-            process.StartInfo.FileName = @"C:\Program Files (x86)\Graphviz2.38\bin\" + executable;
+            process.StartInfo.FileName = execPath;
             process.StartInfo.Arguments = $"-o\"{outFile}\" -Tpdf \"{dotFile}\"";
             process.Start();
             process.WaitForExit();
@@ -69,6 +76,22 @@ namespace SharpViz
             Console.Read();
 
             return null;
+        }
+
+        private static string GetFilePath(string executable)
+        {
+            switch(OS()) {
+                case PlatformID.Unix:
+                case PlatformID.MacOSX:
+                    return String.Format("/usr/local/bin/{0}", executable);
+                default:
+                    return String.Format(@"C:\Program Files (x86)\Graphviz2.38\bin\{0}.exe", executable);
+            }
+        }
+
+        private static PlatformID OS()
+        {
+            return Environment.OSVersion.Platform;
         }
     }
 }
